add:マイグレーションファイル追加,シーダー追加

+++ database/migrations/2020_12_18_174532_create_m_users_classifications_table.php
@@ -15,12 +15,10 @@ class CreateMUsersClassificationsTable extends Migration
     {
         Schema::create('m_users_classifications', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('user_id')->unsigned()->index();
+            $table->bigInteger('user_id')->unsigned()->index();
             $table->timestamps();
 
-            $table->foreign('user_id')
-            ->references('id')->on('users')
-            ->onDelete('cascade');
+            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
         });
     }
